+++ src/main/java/org/prgrms/kdtspringdemo/customer/controller/CustomerWebController.java
@@ -1,8 +1,8 @@
 package org.prgrms.kdtspringdemo.customer.controller;
 
 import org.prgrms.kdtspringdemo.customer.domain.Customer;
-import org.prgrms.kdtspringdemo.customer.domain.dto.CustomerRequestDto;
-import org.prgrms.kdtspringdemo.customer.domain.dto.CustomerViewDto;
+import org.prgrms.kdtspringdemo.dto.CustomerRequestDto;
+import org.prgrms.kdtspringdemo.dto.CustomerViewDto;
 import org.prgrms.kdtspringdemo.customer.service.CustomerService;
 import org.prgrms.kdtspringdemo.wallet.service.WalletService;
 import org.springframework.ui.Model;
