+++ src/main/java/org/prgrms/kdtspringdemo/dto/VoucherRequestDto.java
@@ -1,4 +1,4 @@
-package org.prgrms.kdtspringdemo.voucher.domain.dto;
+package org.prgrms.kdtspringdemo.dto;
 
 public class VoucherRequestDto {
     private String voucherPolicy;
