+++ src/main/java/org/prgrms/kdtspringdemo/wallet/controller/WalletController.java
@@ -3,9 +3,9 @@ package org.prgrms.kdtspringdemo.wallet.controller;
 import org.prgrms.kdtspringdemo.view.InputConsole;
 import org.prgrms.kdtspringdemo.view.OutputConsole;
 import org.prgrms.kdtspringdemo.voucher.domain.Voucher;
-import org.prgrms.kdtspringdemo.voucher.domain.dto.VoucherViewDto;
+import org.prgrms.kdtspringdemo.dto.VoucherViewDto;
 import org.prgrms.kdtspringdemo.wallet.domain.Wallet;
-import org.prgrms.kdtspringdemo.wallet.domain.dto.WalletViewDto;
+import org.prgrms.kdtspringdemo.dto.WalletViewDto;
 import org.prgrms.kdtspringdemo.wallet.service.WalletService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
